+++ optics-vl/CHANGELOG.md
@@ -1,5 +1,7 @@
 # optics-vl-0.3 (2026-??-??)
 * Drop support for GHC older than 9.2.
+* Re-export `GetterVL`, `getterVL` and `toGetterVL` as well as `FoldVL`,
+  `foldVL` and `toFoldVL` from `optics-core`.
 
 # optics-vl-0.2.1 (2020-04-15)
 * Provide conversions from `Iso`/`Prism` to their VL representation